Move global state poller to globalState file
