+++ packages/core/src/codewhisperer/region/regionProfileManager.ts
@@ -24,9 +24,7 @@ import { localize } from '../../shared/utilities/vsCodeUtils'
 import { IAuthProvider } from '../util/authUtil'
 import { Commands } from '../../shared/vscode/commands2'
 import { CachedResource } from '../../shared/utilities/resourceCache'
-import { fs } from '../../shared/fs/fs'
-import path from 'path'
-import { globalKey } from '../../shared/globalState'
+import { GlobalStatePoller } from '../../shared/globalState'
 
 // TODO: is there a better way to manage all endpoint strings in one place?
 export const defaultServiceConfig: CodeWhispererConfig = {
@@ -60,7 +58,6 @@ export class RegionProfileManager {
     public readonly onDidChangeRegionProfile = this._onDidChangeRegionProfile.event
     // Store the last API results (for UI propuse) so we don't need to call service again if doesn't require "latest" result
     private _profiles: RegionProfile[] = []
-    //private sharedState: SharedState
 
     private readonly cache = new (class extends CachedResource<RegionProfile[]> {
         constructor(private readonly profileProvider: () => Promise<RegionProfile[]>) {
@@ -127,18 +124,14 @@ export class RegionProfileManager {
     constructor(private readonly authProvider: IAuthProvider) {
         const getProfileFunction = () =>
             globals.globalState.tryGet<{ [label: string]: RegionProfile }>('aws.amazonq.regionProfiles', Object, {})
-
         const profileChangedHandler = async () => {
             const profile = await this.loadPersistedRegionProfle()
             void this._switchRegionProfile(profile[this.authProvider.profileName], 'reload')
         }
 
+        // This is a poller that handles synchornization of selected region profiles between different IDE windows.
+        // It checks for changes in global state of region profile, invoking the change handler to switch profiles
         GlobalStatePoller.create(getProfileFunction, profileChangedHandler)
-
-        //     , async () => {
-        //     void this._switchRegionProfile((await this.loadPersistedRegionProfle())[this.authProvider.profileName], 'user')
-        // })
-        //this.sharedState = new SharedState('aws.amazonq.regionProfiles')
     }
 
     async getProfiles(): Promise<RegionProfile[]> {
@@ -269,8 +262,6 @@ export class RegionProfileManager {
         // persist to state
         await this.persistSelectRegionProfile()
 
-        //this.sharedState.updateSharedState(this._activeRegionProfile)
-
         // Force status bar to reflect this change in state
         await Commands.tryExecute('aws.amazonq.refreshStatusBar')
     }
@@ -342,13 +333,6 @@ export class RegionProfileManager {
 
         previousPersistedState[this.authProvider.profileName] = this.activeRegionProfile
         await globals.globalState.update('aws.amazonq.regionProfiles', previousPersistedState)
-
-        const persistedState = globals.globalState.tryGet<{ [label: string]: RegionProfile }>(
-            'aws.amazonq.regionProfiles',
-            Object,
-            {}
-        )
-        console.log(persistedState)
     }
 
     async generateQuickPickItem(): Promise<DataQuickPickItem<string>[]> {
@@ -392,7 +376,6 @@ export class RegionProfileManager {
         if (arn) {
             if (this.activeRegionProfile && this.activeRegionProfile.arn === arn) {
                 this._activeRegionProfile = undefined
-                //this.sharedState.updateSharedState(this._activeRegionProfile)
             }
 
             const profiles = this.loadPersistedRegionProfle()
@@ -450,84 +433,5 @@ export class RegionProfileManager {
         )) as CodeWhispererUserClient
 
         return c
-    }
-}
-
-export class GlobalStatePoller {
-    protected oldValue: any
-    protected getState: Function
-    protected changeHandler: Function
-
-    constructor(getState: Function, changeHandler: Function) {
-        this.getState = getState
-        this.changeHandler = changeHandler
-        this.oldValue = getState()
-    }
-
-    static create(getState: Function, changeHandler: () => void) {
-        const instance = new GlobalStatePoller(getState, changeHandler)
-        instance.poll()
-        return instance
-    }
-
-    poll() {
-        const interval = 1000 // ms
-        setInterval(() => {
-            const newValue = this.getState()
-            if (this.oldValue !== newValue) {
-                this.oldValue = newValue
-                this.changeHandler()
-            }
-        }, interval)
-    }
-}
-
-export class SharedState {
-    readonly key: globalKey
-    protected value: any
-    readonly filePath: string
-
-    constructor(key: globalKey) {
-        this.key = key
-        this.filePath = path.join(globals.context.globalStorageUri.fsPath, `${key}.json`)
-    }
-
-    static async create(key: globalKey, changeHandler: () => void) {
-        const instance = new SharedState(key)
-        await instance.createGlobalStateFile()
-        const fileWatcher = instance.getSharedStateFileWatcher()
-        fileWatcher.onDidChange(changeHandler)
-    }
-
-    protected async createGlobalStateFile() {
-        if (!(await fs.existsFile(this.filePath))) {
-            await this.updateFile()
-        }
-    }
-
-    protected async updateFile() {
-        await fs.writeFile(this.filePath, `${new Date().toISOString()}`, { encoding: 'utf8' })
-    }
-
-    protected getSharedStateFileWatcher() {
-        const watcher = vscode.workspace.createFileSystemWatcher(
-            new vscode.RelativePattern(globals.context.globalStorageUri, `${this.key}.json`)
-        )
-        globals.context.subscriptions.push(watcher)
-        return watcher
-    }
-
-    async updateSharedState(value: any) {
-        const persistedState = globals.globalState.tryGet<{ [label: string]: RegionProfile }>(
-            'aws.amazonq.regionProfiles',
-            Object,
-            {}
-        )
-        console.log(persistedState)
-        if (value === this.value && value !== undefined) {
-            return
-        }
-        this.value = value
-        await this.updateFile()
     }
 }
