+++ packages/core/src/codewhisperer/util/authUtil.ts
@@ -90,7 +90,7 @@ export class AuthUtil implements IAuthProvider {
         this.regionProfileManager.onDidChangeRegionProfile(async () => {
             await this.setVscodeContextProps()
         })
-        lspAuth.registerCacheWatcher((event: string) => this.cacheChangedHandler(event))
+        lspAuth.registerCacheWatcher(async (event: string) => await this.cacheChangedHandler(event))
     }
 
     // Do NOT use this in production code, only used for testing
@@ -277,11 +277,11 @@ export class AuthUtil implements IAuthProvider {
         })
     }
 
-    private cacheChangedHandler(event: string) {
+    private async cacheChangedHandler(event: string) {
         if (event === 'delete') {
-            this.logout()
+            await this.logout()
         } else if (event === 'create') {
-            this.restore()
+            await this.restore()
         }
     }
